Strip data IO dependencies from the container
The location and access methods for data will change over time. Only the
essence of transforming input data into results is to be captured
with the container environment. The error surface due to changes in the
evolving data infrastructure will be minimized by provisioning the
data outside and before the container environment is invoked.