Switch figure output for to PNG (from PDF)
The PDF outputs where found to be not 100% bit-identical on a rerun
-- despite exactly identical rendering data. This could be due to JPG
compression of raster images in the PDF. Switching to PNG seems to
address this problem. Given that the purpose of this dataset is to
document reproducibility, less than yielding publication ready
images, this seems to be a sensible change.